move dummy data out of the component to a dedicated dummy data file remove all logs for a cleaner code
