+++ src/components/authentication/Login.tsx
@@ -32,7 +32,6 @@ const Login = () => {
   });
 
   const onSubmit: SubmitHandler<FormData> = (data) => {
-    console.log(data);
     navigate('/');
   };
